backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.24.1
python-dotenv==1.0.0
aiofiles==23.2.1 
ffmpeg==1.4.0
supabase==2.0.2